fixes #75: remove cast from long to int for the returned tiles count/total fields in REST GET call

+++ geowebcache/core/src/main/java/org/geowebcache/seed/TileBreeder.java
@@ -315,9 +315,9 @@ public class TileBreeder implements ApplicationContextAware {
             Entry<Long, GWCTask> entry = iter.next();
             GWCTask task = entry.getValue();
 
-            ret[idx][0] = (int) task.getTilesDone();
+            ret[idx][0] = task.getTilesDone();
 
-            ret[idx][1] = (int) task.getTilesTotal();
+            ret[idx][1] = task.getTilesTotal();
 
             ret[idx][2] = task.getTimeRemaining();
